chore(binding): move types around
