+++ src/runtime/binding/observation.ts
@@ -1,4 +1,4 @@
-import { ICallable, IIndexable } from '../../kernel/interfaces';
+import { ICallable, IIndexable, IDisposable } from '../../kernel/interfaces';
 import { IScope } from './binding-context';
 import { BindingFlags } from './binding';
 
@@ -43,3 +43,127 @@ export type AccessorOrObserver = IAccessor | IBindingTargetAccessor | IBindingTa
 export interface IObservable<T = any> {
   $observers: Record<string, AccessorOrObserver>;
 }
+
+
+/*
+*  Property observation
+*/
+
+export interface IImmediatePropertySubscriber {
+  (newValue: any, oldValue?: any): void;
+}
+
+export interface IBatchedPropertySubscriber {
+  (newValue: any, oldValue?: any): void;
+}
+
+export interface IImmediatePropertySubscriberCollection {
+  /*@internal*/immediateSubscriber0: IImmediatePropertySubscriber;
+  /*@internal*/immediateSubscriber1: IImmediatePropertySubscriber;
+  /*@internal*/immediateSubscribers: Array<IImmediatePropertySubscriber>;
+  /*@internal*/immediateSubscriberCount: number;
+  /*@internal*/notifyImmediate(newValue: any, previousValue?: any): void;
+  subscribeImmediate(subscriber: IImmediatePropertySubscriber): void;
+  unsubscribeImmediate(subscriber: IImmediatePropertySubscriber): void;
+}
+
+export interface IBatchedPropertySubscriberCollection {
+  /*@internal*/batchedSubscriber0: IBatchedPropertySubscriber;
+  /*@internal*/batchedSubscriber1: IBatchedPropertySubscriber;
+  /*@internal*/batchedSubscribers: Array<IBatchedPropertySubscriber>;
+  /*@internal*/batchedSubscriberCount: number;
+  /*@internal*/notifyBatched(newValue: any, oldValue?: any): void;
+  subscribeBatched(subscriber: IBatchedPropertySubscriber): void;
+  unsubscribeBatched(subscriber: IBatchedPropertySubscriber): void;
+}
+
+export interface IPropertyObserver<TObj extends Object, TProp extends keyof TObj> extends IDisposable, IImmediatePropertySubscriberCollection, IBatchedPropertySubscriberCollection {
+  /*@internal*/observing: boolean;
+  /*@internal*/obj: TObj;
+  /*@internal*/propertyKey: TProp;
+  /*@internal*/ownPropertyDescriptor: PropertyDescriptor;
+  /*@internal*/oldValue?: any;
+  /*@internal*/previousValue?: any;
+  /*@internal*/currentValue: any;
+  /*@internal*/hasChanges: boolean;
+  flushChanges(): void;
+  getValue(): any;
+  setValue(newValue: any): void;
+}
+
+export type PropertyObserver = IPropertyObserver<any, PropertyKey>;
+
+
+
+/*
+*  Collection observation
+*/
+
+export interface IObservedCollection {
+  $observer: CollectionObserver;
+}
+export interface IObservedArray extends IObservedCollection, Array<any> { }
+export interface IObservedSet extends IObservedCollection, Set<any> { }
+export interface IObservedMap extends IObservedCollection, Map<any, any> { }
+export type ObservedCollection = IObservedArray | IObservedSet | IObservedMap;
+
+export const enum CollectionKind {
+  indexed = 0b1000,
+  keyed   = 0b0100,
+  array   = 0b1001,
+  map     = 0b0110,
+  set     = 0b0111
+}
+
+export interface IImmediateCollectionSubscriber {
+  (origin: string, args?: IArguments): void;
+}
+
+export interface IBatchedCollectionSubscriber {
+  (indexMap: Array<number>): void;
+}
+
+export interface IImmediateCollectionSubscriberCollection {
+  /*@internal*/immediateSubscriber0: IImmediateCollectionSubscriber;
+  /*@internal*/immediateSubscriber1: IImmediateCollectionSubscriber;
+  /*@internal*/immediateSubscribers: Array<IImmediateCollectionSubscriber>;
+  /*@internal*/immediateSubscriberCount: number;
+  /*@internal*/notifyImmediate(origin: string, args?: IArguments): void;
+  subscribeImmediate(subscriber: IImmediateCollectionSubscriber): void;
+  unsubscribeImmediate(subscriber: IImmediateCollectionSubscriber): void;
+}
+
+export interface IBatchedCollectionSubscriberCollection {
+  /*@internal*/batchedSubscriber0: IBatchedCollectionSubscriber;
+  /*@internal*/batchedSubscriber1: IBatchedCollectionSubscriber;
+  /*@internal*/batchedSubscribers: Array<IBatchedCollectionSubscriber>;
+  /*@internal*/batchedSubscriberCount: number;
+  /*@internal*/notifyBatched(indexMap: Array<number>): void;
+  subscribeBatched(subscriber: IBatchedCollectionSubscriber): void;
+  unsubscribeBatched(subscriber: IBatchedCollectionSubscriber): void;
+}
+
+export interface ICollectionObserver<T extends ObservedCollection> extends IDisposable, IImmediateCollectionSubscriberCollection, IBatchedCollectionSubscriberCollection {
+  /*@internal*/collection: T;
+  /*@internal*/indexMap: Array<number>;
+  /*@internal*/hasChanges: boolean;
+  /*@internal*/resetIndexMap(): void;
+  flushChanges(): void;
+}
+
+export interface IArrayObserver extends ICollectionObserver<IObservedArray> {
+  lengthPropertyName: 'length';
+  collectionKind: CollectionKind.array;
+}
+
+export interface IMapObserver extends ICollectionObserver<IObservedMap> {
+  lengthPropertyName: 'size';
+  collectionKind: CollectionKind.map;
+}
+
+export interface ISetObserver extends ICollectionObserver<IObservedSet> {
+  lengthPropertyName: 'size';
+  collectionKind: CollectionKind.set;
+}
+
+export type CollectionObserver = IArrayObserver | IMapObserver | ISetObserver;
