+++ src/runtime/binding/observers/collection-observer.ts
@@ -1,74 +1,5 @@
-import { IDisposable } from "../../../kernel/interfaces";
+import { CollectionObserver, IImmediateCollectionSubscriber, IBatchedCollectionSubscriber, IArrayObserver, IMapObserver, CollectionKind, ISetObserver } from "../observation";
 
-
-export interface IObservedCollection {
-  $observer: CollectionObserver;
-}
-export interface IObservedArray extends IObservedCollection, Array<any> { }
-export interface IObservedSet extends IObservedCollection, Set<any> { }
-export interface IObservedMap extends IObservedCollection, Map<any, any> { }
-export type ObservedCollection = IObservedArray | IObservedSet | IObservedMap;
-
-export const enum CollectionKind {
-  indexed = 0b1000,
-  keyed   = 0b0100,
-  array   = 0b1001,
-  map     = 0b0110,
-  set     = 0b0111
-}
-
-export interface IImmediateCollectionSubscriber {
-  (origin: string, args?: IArguments): void;
-}
-
-export interface IBatchedCollectionSubscriber {
-  (indexMap: Array<number>): void;
-}
-
-export interface IImmediateCollectionSubscriberCollection {
-  immediateSubscriber0: IImmediateCollectionSubscriber;
-  immediateSubscriber1: IImmediateCollectionSubscriber;
-  immediateSubscribers: Array<IImmediateCollectionSubscriber>;
-  immediateSubscriberCount: number;
-  notifyImmediate(origin: string, args?: IArguments): void;
-  subscribeImmediate(subscriber: IImmediateCollectionSubscriber): void;
-  unsubscribeImmediate(subscriber: IImmediateCollectionSubscriber): void;
-}
-
-export interface IBatchedCollectionSubscriberCollection {
-  batchedSubscriber0: IBatchedCollectionSubscriber;
-  batchedSubscriber1: IBatchedCollectionSubscriber;
-  batchedSubscribers: Array<IBatchedCollectionSubscriber>;
-  batchedSubscriberCount: number;
-  notifyBatched(indexMap: Array<number>): void;
-  subscribeBatched(subscriber: IBatchedCollectionSubscriber): void;
-  unsubscribeBatched(subscriber: IBatchedCollectionSubscriber): void;
-}
-
-export interface ICollectionObserver<T extends ObservedCollection> extends IDisposable, IImmediateCollectionSubscriberCollection, IBatchedCollectionSubscriberCollection {
-  collection: T;
-  indexMap: Array<number>;
-  hasChanges: boolean;
-  resetIndexMap(): void;
-  flushChanges(): void;
-}
-
-export interface IArrayObserver extends ICollectionObserver<IObservedArray> {
-  lengthPropertyName: 'length';
-  collectionKind: CollectionKind.array;
-}
-
-export interface IMapObserver extends ICollectionObserver<IObservedMap> {
-  lengthPropertyName: 'size';
-  collectionKind: CollectionKind.map;
-}
-
-export interface ISetObserver extends ICollectionObserver<IObservedSet> {
-  lengthPropertyName: 'size';
-  collectionKind: CollectionKind.set;
-}
-
-export type CollectionObserver = IArrayObserver | IMapObserver | ISetObserver;
 
 function notifyImmediate(this: CollectionObserver, origin: string, args?: IArguments): void {
   this.hasChanges = true;
